catalogue/seed/requirements.txt
pydantic
python-slugify